Coin Mechanism Improvements
- The bug disappears somehow?? I didn't do anything and it just disappears when I tried it lmao
- So i just refactored the coin mechanism into a separate class. Much cleaner and easier to read :D
