+++ Assets/Scripts/Utilities/CoinMechanism.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a68e920fb67ef4a42b68dd02e96ef647
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
