+++ Assets/Scripts/Utilities/GameController.cs
@@ -6,25 +6,26 @@ using UnityEngine.UI;
 using UnityEngine.SceneManagement;
 using UnityEngine.Audio;
 
+enum PlayerDataKey
+{
+    CoinAmount,
+    Character,
+}
+
 // GameController is our Singleton Class that serves as the game's controller (duh)
 // It stores the game's global variable such as game currency and player's current skin (future development)
 public class GameController : MonoBehaviour
 {
     public static GameController Instance;
-    private int totalCoin;
     private int currentCharacter; // currentSkin is a variable that holds current characterId
-    public GameObject[] coinAmount;
-    public GameObject[] coinChangeIndicator;
+    public CoinMechanism coinMechanism;
+    [SerializeField] public AudioMixer masterMixer;
+
+    // Scene states variable
     public int selectedLevel = 0;
     public int currentStage = 0; 
     public SceneStateManager.SceneState sceneState = SceneStateManager.SceneState.Onboarding;
-    [SerializeField] public AudioMixer masterMixer;
 
-    private enum PlayerDataKey
-    {
-        CoinAmount,
-        Character,
-    }
 
     private void Awake()
     {
@@ -67,8 +68,7 @@ public class GameController : MonoBehaviour
     {
         // Both currentSkin and totalCoin default value is 0
         this.currentCharacter = PlayerPrefs.GetInt(PlayerDataKey.Character.ToString(), 0);
-        this.totalCoin = PlayerPrefs.GetInt(PlayerDataKey.CoinAmount.ToString(), 0);
-        ShowCoinAmount();
+        coinMechanism = gameObject.AddComponent<CoinMechanism>();
     }
 
     public void ResetMixer()
@@ -78,50 +78,7 @@ public class GameController : MonoBehaviour
         masterMixer.SetFloat("musicVolume", Mathf.Log10(PlayerPrefs.GetFloat(SettingsList.Music.ToString())) * 20);
     }
 
-    void ShowCoinAmount()
-    {
-        // setting the value into the game object 
-        coinAmount = GameObject.FindGameObjectsWithTag("CoinAmount");
-
-        foreach (GameObject c in coinAmount)
-        {
-            c.GetComponent<TMP_Text>().text = totalCoin.ToString();
-        }
-    }
-
-    // animate the decrease or increase in coinAmount
-    // parameter:
-    // - string sign => a "+" or "-"
-    // - int amount => an integer that represents the amount increased or decreased
-    private void AnimateCoinChange(string sign, int amount)
-    {
-        ShowCoinAmount();
-        coinChangeIndicator = GameObject.FindGameObjectsWithTag("CoinChangeIndicator");
-        foreach (GameObject c in coinChangeIndicator)
-        {
-            c.GetComponent<TMP_Text>().text = sign +""+amount.ToString();
-            AnimationUtilities.AnimateAddMoney(c); // calls on AnimationUtilities class
-            break;
-        }
-    }
-
-    // public function where you can add any amount of coin
-    public void CoinAdd(int coinAmount)
-    {
-        AnimateCoinChange("+", coinAmount);
-        this.totalCoin += coinAmount;
-        PlayerPrefs.SetInt(PlayerDataKey.CoinAmount.ToString(), totalCoin); // Automatically saves the new value to PlayerPrefs
-        AudioController.Instance.PlaySound(SoundNames.coinadd);
-    }
-
-    // public function where you can substract any amount of coin
-    public void CoinSubstract(int coinAmount)
-    {
-        AnimateCoinChange("-", coinAmount);
-        this.totalCoin -= coinAmount;
-        PlayerPrefs.SetInt(PlayerDataKey.CoinAmount.ToString(), totalCoin); // Automatically saves the new value to PlayerPrefs
-    }
-
+    
     public void SetCurrentSkin(int skinId)
     {
         this.currentCharacter = skinId;
@@ -131,11 +88,6 @@ public class GameController : MonoBehaviour
     public int GetCurrentSkin()
     {
         return this.currentCharacter;
-    }
-
-    public int GetTotalCoin()
-    {
-        return this.totalCoin;
     }
 
     // a function that detects when the scene has changed
